+++ Ascension/src/blockfighter/server/entities/player/skills/SkillBowStorm.java
@@ -22,7 +22,7 @@ public class SkillBowStorm extends Skill {
     private static final byte SKILL_CODE = Globals.BOW_STORM;
     private static final boolean IS_PASSIVE;
     private static final byte REQ_WEAPON;
-    private static final double MAX_COOLDOWN;
+    private static final long MAX_COOLDOWN;
 
     private static final double BASE_VALUE, MULT_VALUE;
 
@@ -35,7 +35,7 @@ public class SkillBowStorm extends Skill {
         HashMap<String, Integer> dataHeaders = Globals.getDataHeaders(data, CUSTOM_DATA_HEADERS);
 
         REQ_WEAPON = Globals.loadReqWeapon(data, dataHeaders);
-        MAX_COOLDOWN = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MAXCOOLDOWN_HEADER);
+        MAX_COOLDOWN = (long) Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MAXCOOLDOWN_HEADER);
         BASE_VALUE = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_BASEVALUE_HEADER);
         MULT_VALUE = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MULTVALUE_HEADER);
         IS_PASSIVE = Globals.loadBooleanValue(data, dataHeaders, Globals.SKILL_PASSIVE_HEADER);
@@ -67,7 +67,7 @@ public class SkillBowStorm extends Skill {
     }
 
     @Override
-    public double getMaxCooldown() {
+    public long getMaxCooldown() {
         return MAX_COOLDOWN;
     }
