+++ Ascension/src/blockfighter/server/entities/player/skills/SkillPassiveResistance.java
@@ -10,7 +10,7 @@ public class SkillPassiveResistance extends SkillPassive {
 
     private static final boolean IS_PASSIVE;
     private static final byte REQ_WEAPON;
-    private static final double MAX_COOLDOWN;
+    private static final long MAX_COOLDOWN;
 
     private static final double BASE_VALUE, MULT_VALUE;
 
@@ -19,7 +19,7 @@ public class SkillPassiveResistance extends SkillPassive {
         HashMap<String, Integer> dataHeaders = Globals.getDataHeaders(data, null);
 
         REQ_WEAPON = Globals.loadReqWeapon(data, dataHeaders);
-        MAX_COOLDOWN = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MAXCOOLDOWN_HEADER);
+        MAX_COOLDOWN = (long) Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MAXCOOLDOWN_HEADER);
         BASE_VALUE = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_BASEVALUE_HEADER);
         MULT_VALUE = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MULTVALUE_HEADER);
         IS_PASSIVE = Globals.loadBooleanValue(data, dataHeaders, Globals.SKILL_PASSIVE_HEADER);
@@ -55,8 +55,8 @@ public class SkillPassiveResistance extends SkillPassive {
     }
 
     @Override
-    public double getMaxCooldown() {
-        return MAX_COOLDOWN - (BASE_VALUE + MULT_VALUE * this.level);
+    public long getMaxCooldown() {
+        return (long) (MAX_COOLDOWN - (BASE_VALUE + MULT_VALUE * this.level));
     }
 
     @Override
