+++ Ascension/src/blockfighter/server/entities/player/skills/SkillSwordGash.java
@@ -20,7 +20,7 @@ public class SkillSwordGash extends Skill {
     private static final byte SKILL_CODE = Globals.SWORD_GASH;
     private static final boolean IS_PASSIVE;
     private static final byte REQ_WEAPON;
-    private static final double MAX_COOLDOWN;
+    private static final long MAX_COOLDOWN;
     private static final double BASE_VALUE, MULT_VALUE;
     private static final byte REQ_EQUIP_SLOT = Globals.ITEM_WEAPON;
     private static final byte PLAYER_STATE = Player.PLAYER_STATE_SWORD_GASH;
@@ -31,7 +31,7 @@ public class SkillSwordGash extends Skill {
         HashMap<String, Integer> dataHeaders = Globals.getDataHeaders(data, CUSTOM_DATA_HEADERS);
 
         REQ_WEAPON = Globals.loadReqWeapon(data, dataHeaders);
-        MAX_COOLDOWN = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MAXCOOLDOWN_HEADER);
+        MAX_COOLDOWN = (long) Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MAXCOOLDOWN_HEADER);
         BASE_VALUE = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_BASEVALUE_HEADER);
         MULT_VALUE = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MULTVALUE_HEADER);
         IS_PASSIVE = Globals.loadBooleanValue(data, dataHeaders, Globals.SKILL_PASSIVE_HEADER);
@@ -58,7 +58,7 @@ public class SkillSwordGash extends Skill {
     }
 
     @Override
-    public double getMaxCooldown() {
+    public long getMaxCooldown() {
         return MAX_COOLDOWN;
     }
